--- src/main/kotlin/backend/itracker/crawl/service/macbook/MacBookCrawlMapper.kt
@@ -1,4 +1,0 @@
-package backend.itracker.crawl.service.macbook
-
-class MacBookCrawlMapper {
-}
